[eRCP] Remove obsolete JFaceSupport class
The getJFaceFonts() method has been moved to the JFaceFontPage.
